Tensorflow/YOLO/requirements.txt
tensorflow == 2.10.0
Pillow == 10.0.0
seaborn
opencv-python == 4.7.0.68